requirements.txt
asgiref==3.3.4
Django==3.2.2
pytz==2021.1
sqlparse==0.4.1
typing-extensions==3.10.0.0
django-nested-inline==0.4.2
django-crispy-forms==1.11.2
